pyproject.toml
[project]
name = "deorz-blog"
version = "0.1.0"
description = "Персональный блог Denis Orzhenovskiy (deorz)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi == 0.116.0",
    "pydantic-settings == 2.10.1",
    "uvicorn == 0.35.0",
]

[dependency-groups]
dev = [
    "pytest == 8.4.1",
    "pytest-cov == 7.0.0",
    "pytest-mock == 3.14.1",
    "ruff == 0.12.2",
    "syrupy == 4.9.1",
    "ty>=0.0.1a14",
    "httpx == 0.28.1",
    "freezegun == 1.5.5",
]

[tool.pytest.ini_options]
testpaths = ['tests']
